it works but doenst look good
